Add the final files necessary to support IP multicasting.
NOTE:  Due to the way BSDI has dealt with the IP_TOS stuff in netinet/in.h,
it is not possible to preserve our existing values and still be BSDI/NetBSD
compatible for those utilities that care about such things, like vat.
Since Van Jacobsen is not distributing sources for vat, we have no choice
but to use the BSDI one if we want to get in on the fun, and so compatibility
is an issue.  What this means is that people wishing to take advantage of
multicast functionality and also be able to use BSDI/NetBSD binaries will
have to do two things:

	1. Recompile a kernel with `option	IP_VAT_COMPAT'
	2. Recompile things like rlogin, ftp, telnetd, rlogind, ftpd, etc
	   since they'll now be broken with the new kernel.

Folks like Amancio, Jim and Paul T. have taken the opposite approach,
electing to take all their pain up-front and leave this on by default,
having instead an option to turn it *off* (NO_VAT_COMPAT).  I thought
about this and came to the conclusion that a lot of folks would scream
like scalded cats if I did this, so I'm going to leave the multicast
crowd as the odd men out and see about preparing a FAQ entry that describes
how to enable it.

Fair enough?