Don't arbitrarily set SCSI_NOSLEEP.  It is now handled correctly by the higher level scsi code.
Spls should never be conditionalized, so don't do so here.

Restructure the get_ccb routine so that we can't get into an infinite
loop if the ccbs are exhausted and we are are called with SCSI_NOSLEEP set.

Other driver maintainer's that based their ccb allocation routines on Julian's
code should look at these changes and implement them for their driver.

Submitted by: John Dyson


git-svn-id: svn+ssh://svn.freebsd.org/base/head@8031 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f