Yikes!  Originally, i intended to apply the patch from PR # 1322, but Bill has beaten me on this. ;-)
However, he missed the part to compute the kernelsize in kilobytes,
so the minfree consideration was now overcautious.  (I've also
changed the return type of dump_size() to void since int was useless.)

Being here, the fact that `vmcore' was written world-readable was just
a plain security hole: everybody who was able to crash a kernel could
later read any confidential information out of it at his will.  Create
it with umask 077 instead.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@18916 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f