Add device IDs for new Symbios SCSI chips (53c875J,885,895,896), which should work with no driver changes, though not all features are currently used.
Remove code that was conditional on NEW_SCSICONF not being defined. This
was temporary code, that at a time got excluded correctly, until the new
scsiconf became the default, and NEW_SCSICONF was no longer specified.

Add support for quirks defined in scsiconf.c. For now only the HP3724/5
needs an entry, since that drive can't be used with tags.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@26568 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f