Updated the ch(4) driver and chio(1) command to include volume tag support.  These changes have been tested with a Breeze Hill Q47 DLT and a DEC DLT2500 media changer.  The latter has no volume tag support.
The chio(1) command was updated to include various flags to the
status subcommand.  These flags can be used to select additional
information to be displayed (like volume tags).

A new chio(1) subcommand named 'voltag' has been added which allows
for changes to volume tags inside the media changer controller.
This could not be tested as the Q47 does not provide the functio-
nality.

Submitted by:	Hans Huebner


git-svn-id: svn+ssh://svn.freebsd.org/base/head@39227 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f